Added an exception prov.QueriesNotInitedError
This exception is to be raised whenever a query function is called, but the
prov.query_init() function has not been called. This is because it was hurting
my brain every time I saw a KeyError when I knew the key was correct. This error
is far easier to understand.